fix: repair the 32-bit build
GOARCH=386 go vet ./... exited 1: untyped constants wider than a 32-bit int
reached int contexts in ledger, api/utxorpc, config.go, mithril, ouroboros,
internal/nodeparity and two test files.

Guard comparisons widen the operand before comparing, so each guard keeps
its behaviour on a 64-bit target and is well-typed on a 32-bit one. The
offset, length and transaction-index guards in ledger share a new
fitsUint32 helper. mithril's extract-size constants are typed int64, which
they were already compared as.

The 2GiB LocalStateQuery read-buffer cap in internal/nodeparity and
ouroboros is a size rather than a guard; both now read it from
maxLocalStateQueryReadBufferSize, clamped to math.MaxInt where an int
cannot hold 2GiB.

Adds boundary coverage for the guards that had none.

+++ api/utxorpc/query.go
@@ -1123,7 +1123,7 @@ func rationalToUtxorpc(
 }
 
 func uint32FromInt(value int, label string) (uint32, error) {
-	if value < 0 || value > math.MaxUint32 {
+	if value < 0 || int64(value) > math.MaxUint32 {
 		return 0, fmt.Errorf("%s out of range: %d", label, value)
 	}
 	return uint32(value), nil // #nosec G115 -- bounds checked above
